requirements.txt
alembic==0.9.7
click==6.7
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.1
Werkzeug==0.14.1
